Java Core - Cybersoft

+++ Session14_Practice/src/cybersoft/Manager.java
@@ -4,7 +4,7 @@ public class Manager extends Employee {
 
 	private String department;
 
-	public Manager(String name, int age, double salary, String department) {
+	public Manager(String name, int age, int salary, String department) {
 		super(name, age, salary);
 		this.department = department;
 	}
